Compiler warnings (as errors) fix for Clang 10.0 toolchains  - Various issues with implicit double/float to int casts around numeric limits.  - Ranged for loops that take a const copy of the enumerated element, rather than a const reference.
#jira FORT-323870
#rb Jeff.Newquist, Josh.Adams


#ROBOMERGE-OWNER: Luke.Thatcher
#ROBOMERGE-AUTHOR: luke.thatcher
#ROBOMERGE-SOURCE: CL 14529285 via CL 14529295 via CL 14529298
#ROBOMERGE-BOT: (v748-14444958)
#ROBOMERGE-CONFLICT from-shelf
#ROBOMERGE[NEXTGEN]: -Release-4.26

[CL 14529510 by Luke Thatcher in 4.25-Plus branch]

+++ Engine/Source/Runtime/AppFramework/Private/Framework/Testing/STableViewTesting.cpp
@@ -139,7 +139,7 @@ class FTestData
 
 		static float MakeRandomItemHeight()
 		{
-			return static_cast<float>(rand()) / RAND_MAX * 50.0f;
+			return (static_cast<double>(rand()) / RAND_MAX) * 50.0;
 		}
 
 		/**
